+++ src/index.js
@@ -1,7 +1,9 @@
+var Calendar = require('./Calendar');
 var TimeInput = require('./TimeInput');
 var DateInput = require('./DateInput');
 
 module.exports = {
+  Calendar,
   TimeInput,
   DateInput
 };